Completely removed Stripe and the Dropin-UI form has been integrated and positioned accordingly.

+++ app/Http/Controllers/CheckoutController.php
@@ -21,14 +21,44 @@ class CheckoutController extends Controller
             'card-name' => 'required'
         ]);
 
-            Mail::send(new OrderConfirmed);
-
-            Cart::destroy();
-
+        $amount = Cart::total();
+        $quantity = Cart::count(); 
+        $nonce = $request->tokenizationKey;
+        
+        $result = $gateway->transaction()->sale([
+            'amount' => $amount,
+            'quantity' => $quantity,
+            'nonce' => $nonce,
+            'customerDetails' => [
+                'name' => $request->name,
+                'email' => $request->email,
+                'address' => $request->address, 
+                'country' => $request->country,
+                'postcode' => $request->postcode
+            ],
+            'options' => [
+                'submitForSettlement' => true
+            ]
+        ]);
+        
+        if ($result->success || !is_null($result->transaction)) 
+        {
+            $transaction = $result->transaction;
             return redirect()->route('confirmation')
             ->with('paymentSuccessMessage', 'Thank you! Your payment has been accepted.
                                              A confirmation email has also been sent.');
 
+            Mail::send(new OrderConfirmed);
             Cart::destroy(); 
+        } 
+        else 
+        {
+            $errorString = "";
+        
+            foreach($result->errors->deepAll() as $error) 
+            {
+                $errorString .= 'Error: ' . $error->code . ": " . $error->message . "\n";
+            }
+        }
         }
     }
